binlog_fetcher/requirements.txt
mysql-replication
boto3